configuring stripe apikey

+++ backend/src/main/java/com/project/backend/dto/PaymentRequest.java
@@ -1,11 +1,10 @@
 package com.project.backend.dto;
 
-import java.math.BigDecimal;
+
 
 import jakarta.validation.constraints.DecimalMin;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
-import jakarta.validation.constraints.Positive;
 import jakarta.validation.constraints.Size;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -28,7 +27,7 @@ public class PaymentRequest {
 
     @NotNull(message = "Amount is required")
     @DecimalMin(value = "0.50", message = "Minimum amount is 0.50")
-    private BigDecimal amount;
+    private Long amount;
 
     @NotBlank(message = "Currency is required")
     @Size(min = 3, max = 3, message = "Currency must be 3 characters")
